chore: move report model to it's own folder
